+++ lib/screens/auth_screen.dart
@@ -1,0 +1,360 @@
+import 'package:flutter/material.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:google_sign_in/google_sign_in.dart';
+import 'package:photofindapp/screens/photo_search_screen.dart';
+import 'package:photofindapp/services/auth_service.dart';
+
+class AuthScreen extends StatefulWidget {
+  final bool isLogin;
+  const AuthScreen({super.key, required this.isLogin});
+
+  @override
+  AuthScreenState createState() => AuthScreenState();
+}
+
+class AuthScreenState extends State<AuthScreen> {
+  late bool _isLogin;
+  final _formKey = GlobalKey<FormState>();
+  final TextEditingController _emailController = TextEditingController();
+  final TextEditingController _passwordController = TextEditingController();
+  bool _isLoading = false;
+  bool _obscurePassword = true;
+
+  @override
+  void initState() {
+    super.initState();
+    _isLogin = widget.isLogin;
+  }
+
+  @override
+  void dispose() {
+    _emailController.dispose();
+    _passwordController.dispose();
+    super.dispose();
+  }
+
+  Future<void> _signInWithEmail() async {
+    if (_formKey.currentState!.validate()) {
+      setState(() {
+        _isLoading = true;
+      });
+      try {
+        UserCredential userCredential =
+            await FirebaseAuth.instance.signInWithEmailAndPassword(
+          email: _emailController.text.trim(),
+          password: _passwordController.text.trim(),
+        );
+        User? user = userCredential.user;
+        if (user != null) {
+          await AuthService.createUserInFirestore(user);
+          if (mounted) {
+            Navigator.pushReplacement(
+              context,
+              MaterialPageRoute(
+                  builder: (context) => const PhotoSearchScreen()),
+            );
+          }
+        } else {
+          _showSnackBar('Error while logging in');
+        }
+      } on FirebaseAuthException catch (e) {
+        _showSnackBar('Error: ${e.message}');
+      } finally {
+        setState(() {
+          _isLoading = false;
+        });
+      }
+    }
+  }
+
+  Future<void> _signUpWithEmail() async {
+    if (_formKey.currentState!.validate()) {
+      setState(() {
+        _isLoading = true;
+      });
+      try {
+        UserCredential userCredential =
+            await FirebaseAuth.instance.createUserWithEmailAndPassword(
+          email: _emailController.text.trim(),
+          password: _passwordController.text.trim(),
+        );
+        User? user = userCredential.user;
+        if (user != null) {
+          await AuthService.createUserInFirestore(user);
+          if (mounted) {
+            Navigator.pushReplacement(
+              context,
+              MaterialPageRoute(
+                  builder: (context) => const PhotoSearchScreen()),
+            );
+          }
+        } else {
+          _showSnackBar('Error while registering');
+        }
+      } on FirebaseAuthException catch (e) {
+        _showSnackBar('Error: ${e.message}');
+      } finally {
+        setState(() {
+          _isLoading = false;
+        });
+      }
+    }
+  }
+
+  Future<void> _signInWithGoogle() async {
+    setState(() {
+      _isLoading = true;
+    });
+    try {
+      final GoogleSignInAccount? googleUser = await GoogleSignIn().signIn();
+      if (googleUser != null) {
+        final GoogleSignInAuthentication googleAuth =
+            await googleUser.authentication;
+        final credential = GoogleAuthProvider.credential(
+          accessToken: googleAuth.accessToken,
+          idToken: googleAuth.idToken,
+        );
+
+        UserCredential userCredential =
+            await FirebaseAuth.instance.signInWithCredential(credential);
+
+        User? user = userCredential.user;
+        if (user != null) {
+          await AuthService.createUserInFirestore(user);
+          if (mounted) {
+            Navigator.pushReplacement(
+              context,
+              MaterialPageRoute(
+                  builder: (context) => const PhotoSearchScreen()),
+            );
+          }
+        } else {
+          _showSnackBar('Error logging in with Google');
+        }
+      }
+    } on FirebaseAuthException catch (e) {
+      _showSnackBar('Error: ${e.message}');
+    } catch (e) {
+      _showSnackBar('Error: ${e.toString()}');
+    } finally {
+      if (mounted) {
+        setState(() {
+          _isLoading = false;
+        });
+      }
+    }
+  }
+
+  void _showSnackBar(String message) {
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(content: Text(message)),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Stack(
+        children: [
+          Padding(
+            padding: const EdgeInsets.all(16.0),
+            child: _isLoading
+                ? const Center(child: CircularProgressIndicator())
+                : SingleChildScrollView(
+                    child: Form(
+                      key: _formKey,
+                      child: Column(
+                        crossAxisAlignment: CrossAxisAlignment.stretch,
+                        children: [
+                          const SizedBox(height: 80),
+                          _buildLogo(),
+                          const SizedBox(height: 40),
+                          _buildTabSwitcher(),
+                          const SizedBox(height: 20),
+                          _buildEmailField(),
+                          const SizedBox(height: 20),
+                          _buildPasswordField(),
+                          const SizedBox(height: 20),
+                          _buildAuthButton(),
+                          const SizedBox(height: 20),
+                          _buildSocialLoginSeparator(),
+                          const SizedBox(height: 20),
+                          _buildSocialLoginButtons(),
+                          const SizedBox(height: 10),
+                        ],
+                      ),
+                    ),
+                  ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildLogo() {
+    return const Center(
+      child: Icon(
+        Icons.person,
+        size: 60,
+        color: Colors.greenAccent,
+      ),
+    );
+  }
+
+  Widget _buildTabSwitcher() {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: [
+        GestureDetector(
+          onTap: () {
+            setState(() {
+              _isLogin = true;
+            });
+          },
+          child: Text(
+            'Log in',
+            style: TextStyle(
+              fontSize: 20,
+              fontWeight: _isLogin ? FontWeight.bold : FontWeight.normal,
+              color: _isLogin ? Colors.greenAccent : Colors.grey,
+            ),
+          ),
+        ),
+        const SizedBox(width: 20),
+        GestureDetector(
+          onTap: () {
+            setState(() {
+              _isLogin = false;
+            });
+          },
+          child: Text(
+            'Sign up',
+            style: TextStyle(
+              fontSize: 20,
+              fontWeight: !_isLogin ? FontWeight.bold : FontWeight.normal,
+              color: !_isLogin ? Colors.greenAccent : Colors.grey,
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+
+  Widget _buildEmailField() {
+    return TextFormField(
+      controller: _emailController,
+      decoration: InputDecoration(
+        labelText: 'Email',
+        border: OutlineInputBorder(borderRadius: BorderRadius.circular(10)),
+        prefixIcon: const Icon(Icons.email, color: Colors.greenAccent),
+      ),
+      keyboardType: TextInputType.emailAddress,
+      validator: (value) {
+        if (value == null || value.trim().isEmpty) {
+          return 'Please enter your email';
+        }
+        if (!RegExp(r'\S+@\S+\.\S+').hasMatch(value.trim())) {
+          return 'Please enter a valid email address';
+        }
+        return null;
+      },
+    );
+  }
+
+  Widget _buildPasswordField() {
+    return TextFormField(
+      controller: _passwordController,
+      obscureText: _obscurePassword,
+      decoration: InputDecoration(
+        labelText: 'Password',
+        border: OutlineInputBorder(borderRadius: BorderRadius.circular(10)),
+        prefixIcon: const Icon(Icons.lock, color: Colors.greenAccent),
+        suffixIcon: IconButton(
+          icon: Icon(
+            _obscurePassword ? Icons.visibility : Icons.visibility_off,
+            color: Colors.greenAccent,
+          ),
+          onPressed: () {
+            setState(() {
+              _obscurePassword = !_obscurePassword;
+            });
+          },
+        ),
+      ),
+      validator: (value) {
+        if (value == null || value.trim().isEmpty) {
+          return 'Please enter your password';
+        }
+        if (value.trim().length < 6) {
+          return 'Password must be at least 6 characters long';
+        }
+        return null;
+      },
+    );
+  }
+
+  Widget _buildAuthButton() {
+    return ElevatedButton(
+      onPressed: _isLogin ? _signInWithEmail : _signUpWithEmail,
+      style: ElevatedButton.styleFrom(
+        padding: const EdgeInsets.symmetric(vertical: 15),
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
+        backgroundColor: Colors.greenAccent,
+      ),
+      child: Text(_isLogin ? 'Log In' : 'Sign Up'),
+    );
+  }
+
+  Widget _buildSocialLoginSeparator() {
+    return const Row(
+      children: [
+        Expanded(
+          child: Divider(
+            thickness: 1,
+            color: Colors.grey,
+          ),
+        ),
+        Padding(
+          padding: EdgeInsets.symmetric(horizontal: 8.0),
+          child: Text('or'),
+        ),
+        Expanded(
+          child: Divider(
+            thickness: 1,
+            color: Colors.grey,
+          ),
+        ),
+      ],
+    );
+  }
+
+  Widget _buildSocialLoginButtons() {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: [
+        // _buildSocialLoginButton('assets/apple_logo.png', () {}),
+        // const SizedBox(width: 20),
+        _buildSocialLoginButton('assets/google_logo.png', _signInWithGoogle),
+        const SizedBox(width: 20),
+        // _buildSocialLoginButton('assets/facebook_logo.png', () {}),
+        // const SizedBox(width: 20),
+        // _buildSocialLoginButton('assets/phone_icon.png', () {}),
+      ],
+    );
+  }
+
+  Widget _buildSocialLoginButton(String assetPath, VoidCallback onPressed) {
+    return GestureDetector(
+      onTap: onPressed,
+      child: CircleAvatar(
+        radius: 25,
+        backgroundColor: Colors.white,
+        child: Image.asset(
+          assetPath,
+          height: 30,
+          width: 30,
+        ),
+      ),
+    );
+  }
+}
