--- lib/screens/login_screen.dart
@@ -1,281 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:google_sign_in/google_sign_in.dart';
-import 'package:photofindapp/screens/photo_search_screen.dart';
-import 'package:photofindapp/screens/register_screen.dart';
-import 'package:photofindapp/services/auth_service.dart';
-
-class LoginScreen extends StatefulWidget {
-  const LoginScreen({super.key});
-
-  @override
-  LoginScreenState createState() => LoginScreenState();
-}
-
-class LoginScreenState extends State<LoginScreen> {
-  final _formKey = GlobalKey<FormState>();
-  final TextEditingController _emailController = TextEditingController();
-  final TextEditingController _passwordController = TextEditingController();
-  bool _isLoading = false;
-
-  @override
-  void dispose() {
-    _emailController.dispose();
-    _passwordController.dispose();
-    super.dispose();
-  }
-
-  Future<void> _signInWithEmail() async {
-    if (_formKey.currentState!.validate()) {
-      setState(() {
-        _isLoading = true;
-      });
-      try {
-        UserCredential userCredential =
-            await FirebaseAuth.instance.signInWithEmailAndPassword(
-          email: _emailController.text.trim(),
-          password: _passwordController.text.trim(),
-        );
-        User? user = userCredential.user;
-        if (user != null) {
-          await AuthService.createUserInFirestore(user);
-          if (mounted) {
-            Navigator.pushReplacement(
-              context,
-              MaterialPageRoute(
-                  builder: (context) => const PhotoSearchScreen()),
-            );
-          }
-        } else {
-          _showSnackBar('Error al iniciar sesión');
-        }
-      } on FirebaseAuthException catch (e) {
-        _showSnackBar('Error: ${e.message}');
-      } catch (e) {
-        _showSnackBar('Error: ${e.toString()}');
-      } finally {
-        if (mounted) {
-          setState(() {
-            _isLoading = false;
-          });
-        }
-      }
-    }
-  }
-
-  Future<void> _signInWithGoogle() async {
-    setState(() {
-      _isLoading = true;
-    });
-    try {
-      final GoogleSignInAccount? googleUser = await GoogleSignIn().signIn();
-      if (googleUser != null) {
-        final GoogleSignInAuthentication googleAuth =
-            await googleUser.authentication;
-        final credential = GoogleAuthProvider.credential(
-          accessToken: googleAuth.accessToken,
-          idToken: googleAuth.idToken,
-        );
-
-        UserCredential userCredential =
-            await FirebaseAuth.instance.signInWithCredential(credential);
-
-        User? user = userCredential.user;
-        if (user != null) {
-          await AuthService.createUserInFirestore(user);
-          if (mounted) {
-            Navigator.pushReplacement(
-              context,
-              MaterialPageRoute(
-                  builder: (context) => const PhotoSearchScreen()),
-            );
-          }
-        } else {
-          _showSnackBar('Error al iniciar sesión con Google');
-        }
-      }
-    } on FirebaseAuthException catch (e) {
-      _showSnackBar('Error: ${e.message}');
-    } catch (e) {
-      _showSnackBar('Error: ${e.toString()}');
-    } finally {
-      if (mounted) {
-        setState(() {
-          _isLoading = false;
-        });
-      }
-    }
-  }
-
-  void _showSnackBar(String message) {
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(content: Text(message)),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Stack(
-        children: [
-          Container(
-            decoration: const BoxDecoration(
-              image: DecorationImage(
-                image: AssetImage('assets/background.jpg'),
-                fit: BoxFit.cover,
-              ),
-            ),
-          ),
-          Container(
-            color: Colors.blueGrey.withOpacity(0.5),
-          ),
-          Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: _isLoading
-                ? const Center(child: CircularProgressIndicator())
-                : SingleChildScrollView(
-                    child: Form(
-                      key: _formKey,
-                      child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.stretch,
-                        children: [
-                          const SizedBox(height: 100),
-                          _buildLogo(),
-                          const SizedBox(height: 40),
-                          _buildEmailField(),
-                          const SizedBox(height: 20),
-                          _buildPasswordField(),
-                          const SizedBox(height: 20),
-                          _buildSignInButton(),
-                          const SizedBox(height: 10),
-                          _buildGoogleSignInButton(),
-                          const SizedBox(height: 20),
-                          _buildSignUpOption(),
-                        ],
-                      ),
-                    ),
-                  ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildLogo() {
-    return const Center(
-      child: Text(
-        'Photo Find',
-        style: TextStyle(
-          fontSize: 32,
-          fontWeight: FontWeight.bold,
-          color: Colors.white,
-          shadows: [
-            Shadow(
-              blurRadius: 10.0,
-              color: Colors.black45,
-              offset: Offset(3.0, 3.0),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildEmailField() {
-    return TextFormField(
-      controller: _emailController,
-      style: const TextStyle(color: Colors.white),
-      decoration: InputDecoration(
-        labelText: 'Email',
-        labelStyle: const TextStyle(color: Colors.white70),
-        filled: true,
-        fillColor: Colors.black.withOpacity(0.3),
-        border: OutlineInputBorder(borderRadius: BorderRadius.circular(8.0)),
-        prefixIcon: const Icon(Icons.email, color: Colors.white),
-      ),
-      keyboardType: TextInputType.emailAddress,
-      validator: (value) {
-        if (value == null || value.trim().isEmpty) {
-          return 'Please enter your email';
-        }
-        if (!RegExp(r'\S+@\S+\.\S+').hasMatch(value.trim())) {
-          return 'Please enter a valid email address';
-        }
-        return null;
-      },
-    );
-  }
-
-  Widget _buildPasswordField() {
-    return TextFormField(
-      controller: _passwordController,
-      style: const TextStyle(color: Colors.white),
-      decoration: InputDecoration(
-        labelText: 'Password',
-        labelStyle: const TextStyle(color: Colors.white70),
-        filled: true,
-        fillColor: Colors.black.withOpacity(0.3),
-        border: OutlineInputBorder(borderRadius: BorderRadius.circular(8.0)),
-        prefixIcon: const Icon(Icons.lock, color: Colors.white),
-      ),
-      obscureText: true,
-      validator: (value) {
-        if (value == null || value.trim().isEmpty) {
-          return 'Please enter your password';
-        }
-        if (value.trim().length < 6) {
-          return 'Password must be at least 6 characters long';
-        }
-        return null;
-      },
-    );
-  }
-
-  Widget _buildSignInButton() {
-    return ElevatedButton(
-      onPressed: _signInWithEmail,
-      style: ElevatedButton.styleFrom(
-        padding: const EdgeInsets.symmetric(vertical: 15),
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(10),
-        ),
-        backgroundColor: Colors.orange,
-      ),
-      child: const Text('Log In'),
-    );
-  }
-
-  Widget _buildGoogleSignInButton() {
-    return ElevatedButton.icon(
-      onPressed: _signInWithGoogle,
-      icon: Image.asset(
-        'assets/google_logo.png',
-        height: 24,
-      ),
-      label: const Text('Sign in with Google'),
-      style: ElevatedButton.styleFrom(
-        padding: const EdgeInsets.symmetric(vertical: 15),
-        backgroundColor: Colors.white,
-        foregroundColor: Colors.black,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(10),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildSignUpOption() {
-    return TextButton(
-      onPressed: () {
-        Navigator.push(
-          context,
-          MaterialPageRoute(builder: (context) => const RegisterScreen()),
-        );
-      },
-      child: const Text(
-        'Don’t have an account? Sign Up',
-        style: TextStyle(color: Colors.white),
-      ),
-    );
-  }
-}
